# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for EasyFind application.",
        "title": "EasyFind API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/admin/claims/pending": {
            "get": {
                "description": "管理员获取所有待审核的认领申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Claim)"
                ],
                "summary": "管理员获取待审核认领",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/easyfind_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/claims/{id}/approve": {
            "put": {
                "description": "管理员通过认领申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Claim)"
                ],
                "summary": "通过认领申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "认领ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/claims/{id}/reject": {
            "put": {
                "description": "管理员驳回认领申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Claim)"
                ],
                "summary": "驳回认领申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "认领ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/items": {
            "get": {
                "description": "管理员获取所有物品列表，支持筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Item)"
                ],
                "summary": "管理员获取物品列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "校区",
                        "name": "campus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "物品类型 (1:Lost, 2:Found)",
                        "name": "lost_or_found",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/easyfind_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/items/pending": {
            "get": {
                "description": "管理员获取所有待审核的物品列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Item)"
                ],
                "summary": "管理员获取待审核物品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品类型 (1:Lost, 2:Found)",
                        "name": "lost_or_found",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/easyfind_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/items/{id}/approve": {
            "put": {
                "description": "管理员通过物品发布的审核",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Item)"
                ],
                "summary": "通过审核",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/items/{id}/archive": {
            "put": {
                "description": "管理员归档物品帖子",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Item)"
                ],
                "summary": "归档帖子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "处理方式",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.ArchiveRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "归档成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/items/{id}/reject": {
            "put": {
                "description": "管理员驳回物品发布的审核，需提供理由",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin (Item)"
                ],
                "summary": "驳回审核",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "驳回理由",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.RejectRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "easyfind_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "使用 data 统一包裹数据"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.ArchiveRecordRequest": {
            "type": "object",
            "required": [
                "process_method"
            ],
            "properties": {
                "process_method": {
                    "type": "string"
                }
            }
        },
...
        "internal_controllers.RejectRecordRequest": {
            "type": "object",
            "required": [
                "reject_reason"
            ],
            "properties": {
                "reject_reason": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}